rename django_irods module to django_s3
