+++ .coveragerc
@@ -1,7 +1,7 @@
 [run]
 omit =
    /*/python2.7/*
-   django_irods/*
+   django_s3/*
    dublincore/*
    ga_ows/*
    ga_resources/*
