+++ django_s3/exceptions.py
