+++ django_s3/models.py
