+++ django_s3/utils.py
