+++ hs_core/tests/api/native/test_folder_download_zip.py
@@ -6,7 +6,7 @@ from hs_core.hydroshare.users import create_account
 from hs_core.hydroshare.resource import add_resource_files, create_resource
 from hs_core.models import BaseResource
 from hs_core.tasks import create_temp_zip
-from django_irods.storage import S3Storage
+from django_s3.storage import S3Storage
 from hs_core.models import ResourceFile
