+++ local-dev-first-start-only.sh
@@ -315,6 +315,11 @@ docker exec hydroshare chown -R hydro-service:storage-hydro /tmp /shared_tmp
 echo
 
 echo
+echo "  - docker exec -u hydro-service hydroshare python manage.py rename_app django_irods django_s3"
+echo
+docker exec -u hydro-service hydroshare python manage.py rename_app django_irods django_s3
+
+echo
 echo "  - docker exec -u hydro-service hydroshare python manage.py migrate sites --noinput"
 echo
 docker exec -u hydro-service hydroshare python manage.py migrate sites --noinput
